adding support to download models
