+++ script/data/download_dataset.sh
@@ -2,7 +2,12 @@
 
 # path of the dataset folder
 DATASET_NAME=$1
-DATASET_PATH="../../datasets/"
+DATASET_PATH="../../data/datasets/"
+
+# check if the datasets folder exists, if not create it
+if [ ! -d "$DATASET_PATH" ]; then
+    mkdir $DATASET_PATH$
+fi
 
 echo "Downloading dataset ${DATASET_NAME} into ${DATASET_PATH}"
 
@@ -11,7 +16,7 @@ if [ -d "${DATASET_PATH}/${DATASET_NAME}" ]; then
     echo "Dataset folder already exists"
 else
     # execute the python script to download the dataset
-    python3 download_dataset.py $DATASET_NAME
+    python3 download_file.py $DATASET_NAME
 
     # unzip the dataset
     echo "Unzipping dataset"
